+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - login through VK
 - login through Yandex
+- deleting operations
 
 ### Fixed
 - infosnag notifications